gecko_ia2 vbuf __contains__: If accChild fails, try the embedder iframe(s).
In Firefox, when an iframe/frame document is rendered in a separate process to its embedder, accChild for accessibles in the embedded document fails.
If this is the case, we can get the embedder frame and try accChild with the iframe's id.
There can be nested out-of-process frames, so we keep trying this until there are no more frames in the hierarchy (or until we hit the buffer's root id).
We cache frames because walking frames is expensive, and when trying to work out what TreeInterceptor an object belongs to, we'll need to query the frames for each TreeInterceptor.