Use enum for speech reason (PR #10703)
This changes the type of all REASON_* constants.
One outcome of this is that they no longer have equality with
their old string values e.g. REASON_FOCUS == "focus" is False

This would also be the case if the enum used string values instead
of int values.

All occurrences of the previous string values were inspected, and with a
reasonable degree of certainty, there seemed not to be used as REASON_*
values.

To preserve backwards compatibility these enum values are "exploded" to
the module scope. The module level constants should be considered
deprecated and will be removed in the future.