Fix elements list dialog on low resolution screens (PR #10738)
Fixes #10655

With a low screen resolution (800x600), the elements list cuts of some items.

This change reduces the size, and also uses DPI level to scale the tree view control in the elements list dialog.
While doing this I noticed how easy it is to make a mistake with the DpiScalingHelperMixin. I have introduced a new version that does not need to call __init__explicitly in the right order.

In the future, it would be nice to make this dialog resizable, and make this control fill the available space. For now though, this quick approach solves the problem.