speechManager: Ensure handling of skipped indexes (PR #10730)
Fixes #10292

## Summary of the issue:
Synths such as OneCore and sapi5 seem to fail to fire callbacks for bookmarks in speech that directly preceed another bookmark, with no text content in between.
This can happen if doing a sayAll which contains blank lines in the middle.
SpeechManager currently expects that each and every index will be received. But in this case, the index for the blank line is not received, and therefore speech stops.
This was not an issue with 2019.2 and below, presumably because we didn't care so much about indexes for each and every utterance.

## Description of this change
A list of indexes currently sent to the synth is now tracked. 
I.e. the index value of all IndexCommand objects that are in the sequence that is sent to synth.speak are added to the indexesSpeaking instance variable on SpeechManager.
When a an index is reached, do callbacks for skipped indexes as well.