Discard duplicated value for a Windows Explorer list items when accessed via IAccessible (PR #10750)
Summary of the issue:

Windows Explorer list items repeat 'name' as the 'value' property. By default we are accessing it via UIA, and in this case the 'value' property was discarded. There are situations, such as NVDA's own open windows where UIA cannot be used (it causes freezes) and in these cases repetitions were annoying.

The fix:

Similar to what has been done for UIA value of these items is discarded when accessed via IAccessible

Fixes #10620  
Fixes #2395