gecko vbuf backend: When an accessible is moved, do not reuse that node or its descendants (#10776)
* vbufBase: Allow alwaysRerenderDescendants to be set on a node after it is rendered to indicate that we must not reuse descendants in future.

Previously, alwaysRerenderDescendants could be set on a node during rendering and that would prevent any reuse while rendering the subtree.
However, if alwaysRerenderDescendants was set on a node after rendering, it had no effect on future reuse.
Now, it does.

* gecko vbuf backend: When an accessible is moved, do not reuse that node or its descendants.

When an accessible is moved, events are fired as if the accessible were removed and then inserted.
The insertion events are fired as if it were a new subtree; i.e. only one insertion for the root of the subtree.
This means that if new descendants are inserted at the same time as the root is moved, we don't get specific events for those insertions.
Because of that, we mustn't reuse the subtree.
Otherwise, we wouldn't walk inside it and thus wouldn't know about the new descendants.

To achieve this, we set alwaysRerenderDescendants on any node which gets a hide event.
If the node is being removed instead of moved (no subsequent insertion event), this is cheap and simply has no effect.