Fix of error preventing start when log unavailable
Sometimes it occurs that NVDA is unable to create a log file.
This may be caused by no free disk space on the partition
where %temp% folder is located
or missing access privileges.

NVDA should start in such situations anyway.

When log cannot be created, this fix ignores the exception
and continues launching of the program.
When user tries to open a Log Viewer, error window is displayed.