Use commandline config path if provided when opening in explorer (PR #10911)
If NVDA config path is overridden on the command line and the user tries to open configuration directory via globalCommands the wrong profile is opened.

Fix:
globalVars.appArgs.configPath is used to retrieve actual config path falling back to config.getUserDefaultConfigPath() if the argument is not available. 
It may not be available if NVDA_slave is trying to open the directory.