Fix regressions from (PR #10932)
Fixes  #10931
Follow on from #10919

The "report formatting script" (NVDA+f) was causing an error because the `reportSuperscriptsAndSubscripts` key was missing from the `formatConfig` dictionary used as a config section created in ` _reportFormattingHelper`. Updated the generation of this dictionary so this error is unlikely to be encountered again. Instead, now the risk is that any format option separated in the future my not be reported by the "report formatting script" when it should be.

Some superscript/subscript reporting was not separated. To fix this the superscript/subscript fetching has been separated in more situations (PowerPoint, UIA).