oneCore synth: Fix lag between utterances introduced in NVDA 2019.3.
When pushing the last chunk of audio for an utterance, use WavePlayer.sync() instead of WavePlayer.idle().
We do this because idle closes the audio device.
If there's something in the queue after playing the final chunk, that will result in WaveOutOpen being called in the callback when we push the next chunk of audio.
We *really* don't want this because calling WaveOutOpen blocks for ~100 ms if called from the callback when the SSML includes marks, resulting in lag between utterances.
If the queue is empty after sync returns, only then do we call idle and allow the device to close.