Add experimental option to only process IAccessible events for focus ancestry, etc.
Most win events are dropped before they're even added to the limiter unless they're for the focus, a focus ancestor or the desktop object.
This should improve performance in some situations, as well as fixing cases where events for the focus get dropped because of a flood of events for other objects.
Certain events have to be allowed regardless of object; e.g. valueChange for progress bars.
For now, this is an option in advanced settings disabled by default, since there is a risk of unforeseen regressions.
It also breaks eventHandler.requestEvent, though this is also the case for selective UIA event registration.
Nevertheless, it could be useful to have advanced users test this in various situations to see if it resolves problems.
If we decide to pursue this further, we'll need to re-think eventHandler.requestEvent.