gecko_ia2 vbuf backend: Don't unnecessarily calculate labelledByContent (and thus unnecessarily fetch labelledBy).
The labelledByContent ControlField attribute is only used when the name attribute is set and the name is an explicit name.
Previously, this was always calculated, which required the labelledBy relation to be fetched even if we weren't going to use this.
In a document with many objects, this could result in thousands of unnecessary calls.
Instead, only calculate labelledByContent if we're going to use it.