Browser performance:  Don't unnecessarily calculate labelledByContent (PR#11205)
gecko_ia2 vbuf backend: Don't unnecessarily calculate labelledByContent (and thus unnecessarily fetch labelledBy).

The labelledByContent ControlField attribute is only used when the name attribute is set and the name is an explicit name. This is true for how we use the attribute from Python, but it also must be true because of what a browser provides.

When aria-labelledby is provided, the browser uses it to calculate the accessible name and explicit name is also set. This means that accessible name and explicit name (which are already being fetched) can be used as a canary for the existence of the labelledBy relation.

Avoiding calls to fetch the labelledBy relation when it is not going to be used improves performance.
In a document with many objects, this could result in thousands of unnecessary calls.
Instead, only calculate labelledByContent if we're going to use it.