JAB: do not treat ActiveDescendentChange event from not focused component as focus gained event (PR #11048)
Fixes: #5989

When a PropertyActiveDescendentChange event was received, nvda transformed it into a event_gainFocus event.

The issue is that PropertyActiveDescendentChange can be raised by components (e.g. tree view) that don't have the 
focus. In that case, a PropertyActiveDescendentChange does not imply that the newly selected tree node has 
received the focus.

Now, treat PropertyActiveDescendentChange event as 'focus gained event' only if event source has focus or it is an ancestor is the focused component.