Open With: treat Shell-Flyout as a good UIA window (#11336)
* appModules/Open With: copyright header update (note that this file was first added in 2011 as a temporary workaround for Open With issue in Windows 8 preview. Re #11335

* appModules/open with: treat Shell_Flyout as a good UIA window. Re #11335.

ON some systems running Windows 10 May 2020 UPdate (Version 2004), Open With dialog was not read, although NVDA will say 'pane'. Turns out this is a UIA element that is seen as an IAccessible by NVDA. Thus classify Sheell_Flyout window as a good UIA window. NOt only does this let NVDA announce Open With dialog in May 2020 Update, it also lets nVDA not announce 'pane' when Open With opens.

* Update what's new for pr #11336

Co-authored-by: Michael Curran <mick@nvaccess.org>