Synthesizer dialog: ensure 'Microsoft Sound Mapper' is named (PR #11353)
Replace empty string in output device names with 'Microsoft Sound Mapper'.

Fixes #11349.

On Windows 10 Version 2004 (20H1) and 20H2, an empty string is returned when asking Windows for the friendly name of WAVE_MAPPER constant (-1). In older Windows releases, this is named 'Microsoft Sound Mapper'. Therefore call it that when enumerating device names for 'output device' list in synthesizer dialog in NVDA.

Co-authored-by: Reef Turner <reef@nvaccess.org>