NVDAObjects.UIA.UIA: introduce UIAAutomationId getter. Re #11445.
Sometimes a COM error is raised when trying to obtain cached automation Id, particularly if a UIA element isn't ready yet (for example, in modern keyboard). To avoid COM error, introduce a dedicated UIA Automation Id getter that either returns the Automation Id if found or an empty string. The overall structure is based on UIAFullDescription and UIAHelpText getters.