Modern keyboard: rename emoji panel just opened flag to modern keyboard active to better convey what the flag is supposed to do. Re #11454.
When modern keyboard interface opens, a private 'emoji panel just opened' flag is set to inform NVDA that modern keyboard interface is shown. This flag is cleared when a name change event occurs, particularly in version 1809 and later where a name change event is fired when modern keyboard closes. Therefore rename this flag to better reflect what it actually represents.