Modern keyboard: do not clear modern keyboard active flag from name change event unless the object in question no longer has location specified. Re #11454.
In Version 1809 and later, name change event is fired just as emoji panel is being closed. If this happens, object's location becomes undefined (0, 0, 0, 0). Therefore use this as a hint to clear app module flags. Also, do not clear modern keyboard active flag as long as the element for which name change event was fired is shown on screen (location entries are not 0).