Modern keyboard/window open event: navigate to the actual emojis and lcipboard history items.
Rather than stopping at a location where text might be spoken, navigate to actual emoji and clipboard history items. This is seen with:
* Emoji: skin tone modifier is selected when emoji panel first opens to People emoji group. If so, move navigator object to actual emoji item (next object).
* Clipboard history: the object in question may contain an explanatory text or is the actual clipboard history list, not items. If it is the latter, move one more level down.
A combination of these makes life easier for element selected event as it can focus on actual items, not moving to different objects because of traversal errors.