nvwave: Fix occasional exception related to close timer.
The idle close timer must be started on the main thread.
However, it's possible that the player might be closed while the timer start call is waiting to run.
Previously, this would cause an exception when the timer expired because we'd try to close when we were already closed.
To fix this, we now check (on the main thread) if the player is closed, and if it is, we don't start the timer.