Context sensitive help (PR #11456)
Fixes #7757
Fixes #8354
Supersedes PR #8355

Add initial support for opening the user guide at a particular section based on the context when the user presses F1 (in NVDA dialogs). For the sake of simplicity, this is done by opening the user guide in the browser. Unfortunately, this results in a lot of unnecessary verbosity. However, this is still quicker for the user than manually opening the user guide separately (also in a browser) and navigating to the chosen section manually. 

Co-authored-by: ThomasStivers <thomas.stivers@gmail.com>
Co-authored-by: Leonard de Ruijter <leonardder@users.noreply.github.com>