Use dialog instead of Window
Many NVDAObjects derive from Window, including IAccessible.
This meant that the fix for focus changes in Gmail was bypassed!

Looking at the log, after opening NVDA Find dialog the last
queued focus object is set to the following:
- NVDAObjects.Dynamic_NvdaDialogDialogIAccessibleWindowNVDAObject
- NVDAObjects.Dynamic_IAccessibleEditWindowNVDAObject

The first is the dialog, the second the edit field.

The full MRO for the dialog is:
- class 'NVDAObjects.Dynamic_NvdaDialogDialogIAccessibleWindowNVDAObject'
- class 'appModules.nvda.NvdaDialog'
- class 'NVDAObjects.behaviors.Dialog'
- class 'NVDAObjects.IAccessible.IAccessible'
- class 'NVDAObjects.window.Window'
- class 'NVDAObjects.NVDAObject'
- class 'documentBase.TextContainerObject'
- class 'baseObject.ScriptableObject'
- class 'baseObject.AutoPropertyObject'
- class 'garbageHandler.TrackedObject'
- class 'object'

Excluding speech cancellation on 'NVDAObjects.behaviors.Dialog' seems to
generally make sense, resolves the NVDA Find dialog speech for focus
cancellation, and does not appear to interfere with the Gmail fix.