Fix freeze in Chromium with UIA
Trade off: slight degradation of performance.

UIATextInfo._getTextWithFieldsForUIARange:
Test each child to see if it is clipped by or lives completely
outside the parent text range, and if so appropriately clip
the child range and break out of the for loop.

Previously these checks only happened on the final child to try
and avoid extra calls, but Edge sometimes returns many more
children then it should when on the end of a list, and a trivial
slow down is much preferred over a 20 second freeze.