Fix freeze in Chromium with UIA
Trade-off: only the first line is fetched.

speech.getObjectSpeech:
If the given object should have its text presented (E.g. a document
or edit field) and the line at the selection cannot be fetched, then
only present the first line.

Previously the text for the entire object would be fetched, which in
the case of Edgium, was extremely costly and could cause a freeze of
10 seconds or more.