Add braille viewer braille routing (PR #11804)
Pressing buttons on the braille viewer window would not work, since this would change the OS focus. Routing would no longer have the intended effect.

Instead a mechanism has been devised that allows the user to interact with the virtual braille cells without changing focus. The user must instead hover the mouse over a braille cell for a short time. Feedback is given via the background color of the cell. The cell background immediately changes to a new color (yellow) to indicate the timer has started, then transitions gradually to orange, then snaps to green when routing is activated.