Prevent infinite recursion for unspeakable chars
Fixes #11752
When reading characters that do not have any description in NVDA, a 'no content'
utterance can be produced. In this case the empty string is the last
'command' in the sequence passed to 'ensureEndUtterance', resulting in
the sequence not being given an indexCommand or an EndUtteranceCommand.
The indexCommand is required for the cancellable speech processing, to
be able to track which utterance a cancellableSpeechCommand belongs to.