winVersion.WinVersion: add repr method. Re #11837.
Add repr method to be used to print Windows version information to the console and to announce it at startup (future commit). It uses a helper function to obtain release name, then adds version string, service pack (if any), and product type. In essence, the work of old winVersionText was transfered to repr.