Object presentation panel/report notifications: checkbox -> checkable list.
Report notifications is now a checkable list with two items:
* Help balloons and toasts: this partially resurrects the old name for this 'checkbox'.
* App notifications: announce notifications from apps.
Two distinct keys are used to check these boxes. Not only this preserves backward compatibility, it also allows other kinds of notifications to be added later, the first of which will be app notifications (chiefly UIA notification event for now but can be expanded to cover other notifications).