Report 'unsupported' when toggling screen layout in Word (#10795)
Report 'unsupported' when toggling screen layout in Word (PR #10795)

Fixes #7297
Report "unsupported" when trying to toggle screen layout (NVDA+v)
in MS Word and other applications that don't have support for
screen layout.

Previously 'v' was passed through to the application which was 
confusing for users who expected screen layout to be available.

Co-authored-by: Reef Turner <reef@nvaccess.org>