Auto fix generated files for IDE / tools.
Generated "friedly name" comtypes.gen files are modified so that they
can be used by tools / IDEs more easily.

The files are read, the module name extracted, and the import statement
replaced by a more elaborate version which include a fall back to a do a
relative import if the comtypes.gen file is not found.