Make com interfaces ide friendly (PR #12201)
* Make UIAutomationClient work for tools and runtime

At runtime the import from comtypes.gen will work, as a fallback for IDE
/ Tools that can't find that, the relative import is used.

Note: IDE's wont be aware of 'protected' symbols (beginning with
underscore). Protected symbols aren't imported with the from x import * statment.

* Auto fix generated files for IDE / tools.

Generated "friendly name" comtypes.gen files are modified so that they
can be used by tools / IDEs more easily.

The files are read, the module name extracted, and the import statement
replaced by a more elaborate version which include a fall back to a do a
relative import if the comtypes.gen file is not found.