Ensure that `speech.cancelSpeech` clears instance of `SpeechWithoutPauses` which has been used for say all. (#12228)
Fixes #12225

PR #12195 removed speech.speakWithoutPauses which was an alias for speech._speakWithoutPauses. While all usages of speech.speakWithoutPauses were found and fixed it turned out than when cancelling speech during say all speech._speakWithoutPauses was cleared rather than instance of SpeakWithoutPauses used for say all.

Now the instance of SpeechWithoutPauses which is used during say all is cleared. A function is added to sayAllHandler which creates it first if necessary, and ensures that the single instance is being used for say all.